Fixes ruby version changing on cd problems
Turns out that async stuff runs on the original directory if we don't
take a step to `cd` to the current directory every time.

Added a `$PWD` change like the git sections so it will work with stuff
like rbenv that selects the ruby version based on a special file in each
directory.