Adds async version to node
`nvm` works differently than other version managers: it manipulates the
`$PATH` to point at different places, and `nvm current` uses that
`$PATH` directly.

That means that we need to pass the current updated `$PATH` to the async
function every time, or it will use the past value (that never changes).